# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Fellowship Entrance Exam and Seat Allocation System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /units:
    get:
      operationId: listUnits
      tags: [units]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Unit" }
    post:
      operationId: createUnit
      tags: [units]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewUnit"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Unit" }
...
  /preferences/me:
    get:
      operationId: listMyPreferences
      tags: [preferences]
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Preference" }
    put:
      operationId: setMyPreferences
      tags: [preferences]
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPreferencesBody"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Preference" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Unit:
      type: object
      properties:
        id: { type: integer }
        name: { type: string }
        city: { type: string }
        state: { type: string }
        candidateCount: { type: integer }
      required: [id, name, city, state, candidateCount]
    NewUnit:
      type: object
      properties:
        name: { type: string }
        city: { type: string }
        state: { type: string }
      required: [name, city, state]
...
    Preference:
      type: object
      properties:
        id: { type: integer }
        specialityId: { type: integer }
        specialityName: { type: string }
        programName: { type: string }
        preferenceOrder: { type: integer }
      required: [id, specialityId, specialityName, programName, preferenceOrder]
    SetPreferencesBody:
      type: object
      properties:
        specialityIds:
          type: array
          items: { type: integer }
      required: [specialityIds]
...